+++ package.json
@@ -31,14 +31,24 @@
   ],
   "scripts": {
     "build": "bun build ./src/index.ts --outdir ./dist --target browser --format esm --sourcemap && cp ./src/index.d.ts ./dist/index.d.ts",
-    "dev": "bun build ./src/index.ts --watch --outdir ./dist --target browser --format esm --sourcemap"
+    "dev": "bun build ./src/index.ts --watch --outdir ./dist --target browser --format esm --sourcemap",
+    "lint": "eslint . --ext .ts,.tsx,.js,.cjs,.mjs",
+    "lint:fix": "eslint . --ext .ts,.tsx,.js,.cjs,.mjs --fix"
   },
   "dependencies": {
     "html2canvas": "^1.4.1"
   },
   "devDependencies": {
     "@types/bun": "latest",
-    "typescript": "^5.5.0"
+    "@typescript-eslint/eslint-plugin": "^7.0.0",
+    "@typescript-eslint/parser": "^7.0.0",
+    "eslint": "^9.7.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsdoc": "^48.2.12",
+    "eslint-plugin-promise": "^6.4.0",
+    "typescript": "^5.5.0",
+    "typescript-eslint": "^7.0.0"
   },
   "peerDependencies": {}
 }
