+++ src/internal/webgl.ts
@@ -1,0 +1,35 @@
+// WebGL helpers
+
+export function compileShader(gl: WebGLRenderingContext | WebGL2RenderingContext, type: number, src: string) {
+  const shader = gl.createShader(type);
+  if (!shader) return null;
+  gl.shaderSource(shader, src.trim());
+  gl.compileShader(shader);
+  if (!gl.getShaderParameter(shader, gl.COMPILE_STATUS)) {
+    console.error("Shader error", gl.getShaderInfoLog(shader));
+    gl.deleteShader(shader);
+    return null;
+  }
+  return shader;
+}
+
+export function createProgram(
+  gl: WebGLRenderingContext | WebGL2RenderingContext,
+  vsSource: string,
+  fsSource: string
+) {
+  const vs = compileShader(gl, gl.VERTEX_SHADER, vsSource);
+  const fs = compileShader(gl, gl.FRAGMENT_SHADER, fsSource);
+  if (!vs || !fs) return null;
+  const program = gl.createProgram();
+  if (!program) return null;
+  gl.attachShader(program, vs);
+  gl.attachShader(program, fs);
+  gl.linkProgram(program);
+  if (!gl.getProgramParameter(program, gl.LINK_STATUS)) {
+    console.error("Program link error", gl.getProgramInfoLog(program));
+    return null;
+  }
+  return program;
+}
+
